+++ Tools/Colorize.js
@@ -4,14 +4,13 @@
 import { join , normalize , relative , resolve } from 'https://deno.land/std/path/mod.ts'
 import { parse , stringify } from 'https://deno.land/x/xml/mod.ts'
 import { walk , emptyDir , ensureFile , copy } from 'https://deno.land/std/fs/mod.ts'
-import * as YAML from 'https://deno.land/std/encoding/yaml.ts'
 
 import { newline , center , yellow , blue , cyan , green , red } from './Colorizer/Pretty.js'
 import { templatePath } from './Colorizer/Parameter.js'
 import generateCache from './Colorizer/Cache.js'
 import * as Colorize from './Colorizer/ModifySVG.js'
 import * as Path from './Colorizer/Paths.js'
-
+import loadTemplate from './Colorizer/Template/Parse.js'
 
 const
     task1 = blue('① '),
@@ -21,7 +20,6 @@ const
 
 const { readTextFile , writeTextFile , symlink , exit } = Deno;
 const { log , clear } = console;
-
 
 
 clear();
@@ -54,99 +52,20 @@ if(!templatePath){
     exit();
 }
 
-import template from './Colorizer/Template/Defaults.js'
 
 
-function first(string){
-    return string.at(0) ?? '';
-}
 
-function capitalize(string){
-    return first(string).toUpperCase() +
-        string.slice(1).toLowerCase();
-}
-
-try {
-    
-    const yaml = await readTextFile(templatePath);
-    
-    const raw = YAML.parse(yaml);
-    
-    
-    const normalizeComponent = ([ key , value ]) => 
-        [ capitalize(key) , value ];
-    
-    const isRelevantComponent = ([ key ]) => 
-        key in template;
-        
-    const hasData = ([ _ , value ]) =>
-        value;
-    
-    Object
-    .entries(raw)
-    .map(normalizeComponent)
-    .filter(isRelevantComponent)
-    .filter(hasData)
-    .forEach(([ component , properties ]) => {
-        
-        const attributes = template[component];
-        
-        for(let property in properties){
-
-            let value = properties[property];
-            
-            property = capitalize(property);
-            
-            if(!(property in attributes))
-                continue;
-                
-            
-            switch(property){
-            case 'Color':
-            
-                if(/^None$/i.test(value)){
-                    value = null;
-                    break;
-                }
-                
-                if(/^([0-9a-f]{3}){1,2}$/i.test(value))
-                    break;
-                    
-                throw `「 ${ component } » ${ property } 」 '${ value }' is not a hex color string / 'None'\n`;
-            case 'Alpha':
-                
-                if(typeof value !== 'number')
-                    throw `「 ${ component } » ${ property } 」 '${ value }' is not a float value\n`
-            
-                if(value > 1 || value < 0)
-                    throw `「 ${ component } » ${ property } 」 '${ value }' does not fulfill 0 ≤ Alpha ≤ 1\n`
-            
-                break;
-            }
-            
-            attributes[property] = value;
-        }
-    });
-    
-} catch (error) {
-    
-    printTask = () => {
-        log(red(center(`Couldn't Parse Template`)));
-        newline();
-        newline();
-        log(error);
-    }
-    
-    printScreen();
-    exit();
-}
+const template = await loadTemplate(templatePath);
 
 
 Colorize.init(template);
 
 printTask = () => {
+    
     printProjectFolder();
+    
     newline();
+    
     log(
         blue('① '),
         cyan('Copying Theme')
